1) Module Interface:
    . Added variables 'CellArea' and 'WaterVolume' to the list of arguments on subroutine Modify_Interface3D. 'WaterVolume' is real(8). 

    . Changed the name of one of the arguments in Modify_Interface2D to be 'WaterVolume'

    . Added 'WaterVolume1D' and 'CellArea1D' as local variables in the Module Interface. 
        These variables will be used either by Modify_Interface2D or Modify_Interface3D to 
        unfold the matrix to an array that will be sent to the 0D modules

    . Two new routines were build in the Module Interface to cope with these changes 'UnfoldMatrix2D_R8' and UnfoldMatrix3D_R8
    
2) real(8) was added in all the routines and modules where 'WaterVolume' is used.

3) Minor changes in the Module Bivalve and Interface

4) Module BoxDif: Corrected minor inconsistencies

4)Module Water properties:
    . Added routines for the bivalve output and ModuleBivalve related

git-svn-id: https://mohid.svn.codeplex.com/svn@83908 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Software/MOHIDBase1/ModuleBenthicEcology.F90
@@ -344,7 +344,7 @@ Module ModuleBenthicEcology
         real, pointer, dimension(:  )       :: Temperature
         real, pointer, dimension(:,:)       :: MassInKgFromWater
         real, pointer, dimension(:  )       :: Sediment
-        real, pointer, dimension(:  )       :: WaterVolume
+        real(8), pointer, dimension(:  )       :: WaterVolume
         real, pointer, dimension(:  )       :: CellArea
         real, pointer, dimension(:  )       :: ShortWaveTop
         real, pointer, dimension(:  )       :: ShortWaveAverage
@@ -2862,7 +2862,7 @@ if1 :   if ((ready_ .EQ. IDLE_ERR_     ) .OR.                                 &
         real,    dimension(:  ), pointer, optional  :: UptakePO4w
         real,    dimension(:  ), pointer, optional  :: UptakePO4s
         real,    dimension(:  ), pointer, optional  :: LightFactor
-        real,    dimension(:  ), pointer            :: WaterVolume
+        real(8), dimension(:  ), pointer            :: WaterVolume
         real,    dimension(:  ), pointer            :: CellArea
         integer, dimension(:  ), pointer, optional  :: OpenPoints
         real,    dimension(:  ), pointer, optional  :: ShortWaveAverage
